requirements.txt
torch
arviz
pyro-ppl